+++ ChangeLog.md
@@ -8,6 +8,8 @@ Table:
 
 - Reject table batch request bodies exceeding 4MB.
 - Fix binary table property validation to be 64K bytes not 32K characters.
+- Does not error when table created is a substring of another table.
+- Correctly responds with status 404 on patch with non-existant entity.
 
 ## 2022.04 Version 3.17.1